@@ -7,23 +7,28 @@
         "export": "npm run build && next export"
     },
     "dependencies": {
-        "bootstrap": "^4.2.1",
-        "bulma": "^0.7.2",
+        "await-to-js": "^2.1.1",
+        "babel-plugin-styled-components": "^1.10.0",
+        "bulma": "^0.7.5",
         "date-fns": "^1.29.0",
         "isomorphic-unfetch": "^3.0.0",
-        "next": "7.0.2",
+        "ky": "^0.12.0",
+        "next": "^8.1.0",
         "next-compose-plugins": "^2.1.1",
-        "react": "16.6.3",
-        "react-dom": "16.6.3",
-        "rss-parser": "^3.6.0"
+        "next-ga": "^2.3.4",
+        "react": "^16.8.6",
+        "react-content-loader": "^4.2.2",
+        "react-dom": "^16.8.6",
+        "react-emoji-render": "^1.0.0",
+        "styled-components": "^4.2.0"
     },
     "devDependencies": {
         "@zeit/next-sass": "^1.0.1",
-        "autoprefixer": "^9.3.1",
-        "babel-plugin-transform-define": "^1.3.0",
-        "dotenv": "^6.1.0",
-        "next-images": "^1.0.1",
-        "next-server": "^7.0.2-canary.34",
-        "node-sass": "^4.10.0"
+        "autoprefixer": "^9.5.1",
+        "babel-plugin-transform-define": "^1.3.1",
+        "dotenv": "^7.0.0",
+        "next-images": "^1.1.1",
+        "next-server": "^8.1.0",
+        "node-sass": "^4.12.0"
     }
 }
