@@ -2,5 +2,15 @@ const env = require("./env-config.js");
 
 module.exports = {
     presets: ["next/babel"],
-    plugins: [["transform-define", env]]
+    plugins: [
+        ["transform-define", env],
+        [
+            "styled-components",
+            {
+                ssr: true,
+                displayName: true,
+                preprocess: false
+            }
+        ]
+    ]
 };
