@@ -1,62 +1,82 @@
-import TileErrorBoundary from "./TileErrorBoundary";
-import MarsNewsTile from "../components/MarsNewsTile";
-import CuriosityMissionTile from "../components/CuriosityMissionTile";
-import ScienceDailyMarsNews from "../components/ScienceDailyMarsNews";
-import MarsImage from "../images/mars-landscape.jpg";
+import React from "react";
+import { to } from "await-to-js";
+import NasaNewsCard from "./NasaNewsCard";
+import { MarsFeed } from "../api/feed.api";
+
+const marsFeedApi = new MarsFeed();
+
+// import TileErrorBoundary from "./TileErrorBoundary";
 
 export default function MarsSection() {
+    const [marsNews, setMarsNews] = React.useState([]);
+    const [isMarsNewsLoading, setIsMarsNewsLoading] = React.useState(true);
+    const [curiosityNews, setCuriosityNews] = React.useState([]);
+    const [isCuriosityNewsLoading, setIsCuriosityNewsLoading] = React.useState(
+        true
+    );
+    const [scienceDailyNews, setScienceDailyNews] = React.useState([]);
+    const [
+        isScienceDailyNewsLoading,
+        setIsScienceDailyNewsLoading
+    ] = React.useState(true);
+
+    React.useEffect(() => {
+        async function getMarsNews() {
+            const [error, marsNewsResponse] = await to(marsFeedApi.getNews());
+            if (error) console.log("MARS NEWS ERROR", error);
+            setMarsNews(marsNewsResponse);
+            setIsMarsNewsLoading(false);
+            //console.log("marsNewsResponse", marsNewsResponse);
+        }
+        getMarsNews();
+    }, []);
+
+    React.useEffect(() => {
+        async function getCuriosityMissionUpdates() {
+            const [error, curiosityNewsResponse] = await to(
+                marsFeedApi.getCuriosityMissionUpdate()
+            );
+            if (error) console.log("CURIOSITY NEWS ERROR", error);
+            setCuriosityNews(curiosityNewsResponse);
+            setIsCuriosityNewsLoading(false);
+            //console.log("curiosityNewsResponse", curiosityNewsResponse);
+        }
+        getCuriosityMissionUpdates();
+    }, []);
+
+    React.useEffect(() => {
+        async function getScienceDailyMarsNews() {
+            const [error, scienceDailyNewsResponse] = await to(
+                marsFeedApi.getScienceDailyMarsNews()
+            );
+            if (error) console.log("SCIENCE DAILY NEWS ERROR", error);
+            setScienceDailyNews(scienceDailyNewsResponse);
+            setIsScienceDailyNewsLoading(false);
+            //console.log("scienceDailyNewsResponse", scienceDailyNewsResponse);
+        }
+        getScienceDailyMarsNews();
+    }, []);
+
     return (
-        <section
-            className="section"
-            style={{
-                background: "#eb5757" /* fallback for old browsers */,
-                background:
-                    "-webkit-linear-gradient(to left, #eb5757, #000000)" /* Chrome 10-25, Safari 5.1-6 */,
-                background:
-                    "linear-gradient(to left, #eb5757, #000000)" /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-            }}>
-            <div className="container">
-                <h2 className="title has-text-light">Mars Memo</h2>
-                <div className="tile is-ancestor">
-                    {/* <div className="tile is-12">
-                    <div className="tile is-parent">
-                        <article className="tile is-child notification has-background-grey-dark has-text-light">
-                            <MarsTile {...marsData} />
-                        </article>
-                    </div>
-                </div> */}
-                    <div className="tile is-12">
-                        <div className="tile is-parent">
-                            <TileErrorBoundary>
-                                <MarsNewsTile />
-                            </TileErrorBoundary>
-                        </div>
-                        <div className="tile is-parent">
-                            <TileErrorBoundary>
-                                <CuriosityMissionTile />
-                            </TileErrorBoundary>
-                        </div>
-                        <div className="tile is-parent">
-                            <TileErrorBoundary>
-                                <ScienceDailyMarsNews />
-                            </TileErrorBoundary>
-                        </div>
-                    </div>
-                </div>
-                {/* <div className="tile is-ancestor">
-                <div className="tile is-12">
-                    {marsData.rovers.map((rover, i) => {
-                        return (
-                            <div key={i} className="tile is-parent">
-                                <article className="tile is-child">
-                                    <RoverTile {...rover} />
-                                </article>
-                            </div>
-                        );
-                    })}
-                </div>
-            </div> */}
+        <div className="columns">
+            <div className="column">
+                <NasaNewsCard
+                    isLoading={isMarsNewsLoading}
+                    data={marsNews[0]}
+                />
             </div>
-        </section>
+            <div className="column">
+                <NasaNewsCard
+                    isLoading={isCuriosityNewsLoading}
+                    data={curiosityNews[0]}
+                />
+            </div>
+            <div className="column">
+                <NasaNewsCard
+                    isLoading={isScienceDailyNewsLoading}
+                    data={scienceDailyNews[0]}
+                />
+            </div>
+        </div>
     );
 }
